WebStorage/tests/approved/storage_session_key_null.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_session_key_null</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('sessionStorage' in window) && window.sessionStorage !== null){
			sessionStorage.clear();

			sessionStorage.setItem("name", "user1");
			sessionStorage.setItem("age", "20");
			sessionStorage.setItem("a", "1");
			sessionStorage.setItem("b", "2");

			try {

				test(function() {
					assert_equals(sessionStorage.key(-1), null, "sessionStorage.key(-1)")
				}, "key method test")

				test(function() {
					assert_equals(sessionStorage.key(4), null, "sessionStorage.key(4)")
				}, "key method test")
				
			} catch (e) {
				test(function()
				{
					assert_unreached("key method is failed.");
				}, "key method test")
			}
		} else {
			test(function()
			{
				assert_unreached("sessionStorage is not supported.");
			}, "sessionStorage support test")
		}	
	</script>

 </body>
</html>
